requirements.txt
discord.py==2.3.2
yahoo.fin
requests
pandas
